Terminado (?)

+++ BolsaTrabajo/Controllers/HomeController.cs
@@ -16,14 +16,15 @@ namespace BolsaTrabajo.Controllers
         {
             if (TempData.ContainsKey("Usuario"))
             {
-            var U = TempData["Usuario"] as UsuarioViewModel;
-            TempData["Usuario"] = U;
-            Operacion opBd = new Operacion();
-            var m = new CarrerasViewModel();
-            return View(m);
+                var U = TempData["Usuario"] as UsuarioViewModel;
+                TempData["Usuario"] = U;
+                Operacion opBd = new Operacion();
+                var m = new CarrerasViewModel();
+                m.Tipo = U.Tipo;
+                return View(m);
             }
             return RedirectToAction("Login", "Cuenta");
-          
+
         }
 
 
@@ -32,6 +33,9 @@ namespace BolsaTrabajo.Controllers
         {
             Operacion opBd = new Operacion();
             m.publicacionDetallada = opBd.leerPublicacionDetallada(m.IdPublicacion);
+            var U = TempData["Usuario"] as UsuarioViewModel;
+            TempData["Usuario"] = U;
+            m.Tipo = U.Tipo;
             TempData["Publicacion"] = m;
             return RedirectToAction("Vacantes");
         }
@@ -39,12 +43,21 @@ namespace BolsaTrabajo.Controllers
         [HttpGet]
         public ActionResult Vacantes()
         {
-            if (TempData.ContainsKey("Publicacion"))
+            if (TempData.ContainsKey("Usuario"))
             {
-                var m = TempData["Publicacion"] as CarrerasViewModel;
-                return View(m);
+                var U = TempData["Usuario"] as UsuarioViewModel;
+                TempData["Usuario"] = U;
+
+                if (TempData.ContainsKey("Publicacion"))
+                {
+                    var m = TempData["Publicacion"] as CarrerasViewModel;
+                    m.Tipo = U.Tipo;
+                    return View(m);
+                }
+
+                return RedirectToAction("Index");
             }
-            return RedirectToAction("Index");
+            return RedirectToAction("Login", "Cuenta");
         }
 
         public ActionResult About()
@@ -64,13 +77,14 @@ namespace BolsaTrabajo.Controllers
         [HttpGet]
         public ActionResult Publicacion()
         {
-            var U = new UsuarioViewModel
+            if (TempData.ContainsKey("Usuario"))
             {
-                IdUsuario = 1,
-                Empresa = "Calcetines SA"
-            };
-            TempData["Usuario"] = U;
-            return View();
+                var U = TempData["Usuario"] as UsuarioViewModel;
+                TempData["Usuario"] = U;
+                return View();
+            }
+            return RedirectToAction("Login", "Cuenta");
+
         }
 
         [HttpPost]
@@ -113,38 +127,51 @@ namespace BolsaTrabajo.Controllers
         }
 
         [HttpPost]
-        public ActionResult AdminPublicacion(CarrerasViewModel m , string comand)
+        public ActionResult AdminPublicacion(CarrerasViewModel m, string comand)
         {
-            Operacion opBd = new Operacion();
-            var u = TempData["Usuario"] as UsuarioViewModel;
-            TempData["Usuario"] = u;
-            m.publicacion = opBd.leerPublicacion(m.IdPublicacion);
-            TempData["Publicacion"] = m;
+            if (TempData.ContainsKey("Usuario"))
+            {
+                Operacion opBd = new Operacion();
+                var u = TempData["Usuario"] as UsuarioViewModel;
+                TempData["Usuario"] = u;
+                m.publicacion = opBd.leerPublicacion(m.IdPublicacion);
+                TempData["Publicacion"] = m;
 
-            if (comand.Equals("Eliminar")) {
-                int id = m.IdPublicacion;
-                string sql = String.Format( "Delete from Publicacion where ID = {0};",id);
-                opBd.insertar(sql);
-                sql = String.Format("Delete from Categoria where IdPublicacion = {0};",id);
-                opBd.insertar(sql);
-                return RedirectToAction("AdminPublicacion");
-            } else { 
-           
-            return RedirectToAction("Modificar");
+                if (comand.Equals("Eliminar"))
+                {
+                    int id = m.IdPublicacion;
+                    string sql = String.Format("Delete from Publicacion where ID = {0};", id);
+                    opBd.insertar(sql);
+                    sql = String.Format("Delete from Categoria where IdPublicacion = {0};", id);
+                    opBd.insertar(sql);
+                    return RedirectToAction("AdminPublicacion");
+                }
+                else
+                {
+
+                    return RedirectToAction("Modificar");
+                }
             }
-
+            return RedirectToAction("Login", "Cuenta");
         }
-       
+
 
         [HttpGet]
         public ActionResult Modificar()
         {
-            if(TempData.ContainsKey("Publicacion")) {
-               var m = TempData["Publicacion"] as CarrerasViewModel;
-               TempData["Publicacion"] = m;
-                return View(m);
+            if (TempData.ContainsKey("Usuario"))
+            {
+                var u = TempData["Usuario"] as UsuarioViewModel;
+                TempData["Usuario"] = u;
+                if (TempData.ContainsKey("Publicacion"))
+                {
+                    var m = TempData["Publicacion"] as CarrerasViewModel;
+                    TempData["Publicacion"] = m;
+                    return View(m);
+                }
+                return View();
             }
-            return View();
+            return RedirectToAction("Login", "Cuenta");
         }
         [HttpPost]
         public ActionResult Modificar(CarrerasViewModel m)
@@ -154,14 +181,14 @@ namespace BolsaTrabajo.Controllers
             var mAnterior = TempData["Publicacion"] as CarrerasViewModel;
             m.IdPublicacion = mAnterior.IdPublicacion;
             Operacion opBD = new Operacion();
-           
+
             string sql = String.Format("UPDATE [dbo].[Publicacion]" +
                             "SET [Descricpion] = '{0}',[Vacante] = '{1}',[Requisitos] = '{2}'" +
-                            " WHERE ID = {3}",m.publicacion.Descripcion,m.publicacion.Vacante,m.publicacion.Requisitos,m.IdPublicacion);
+                            " WHERE ID = {3}", m.publicacion.Descripcion, m.publicacion.Vacante, m.publicacion.Requisitos, m.IdPublicacion);
             opBD.insertar(sql);
 
             List<String> listaCarreras = opBD.carreras(m.publicacion.Carreras);
-            opBD.insertar("Delete from Categoria where IdPublicacion = "+m.IdPublicacion+" ;");
+            opBD.insertar("Delete from Categoria where IdPublicacion = " + m.IdPublicacion + " ;");
             foreach (var carr in listaCarreras)
             {
                 sql = String.Format("INSERT INTO [dbo].[Categoria] ([IdPublicacion],[Carrera])" +
